Separate IT launch environment by jdk versions
