+++ agent/pom.xml
@@ -211,22 +211,52 @@
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-surefire-plugin</artifactId>
+                <version>2.19.1</version>
+                <configuration>
+                    <skip>true</skip>
+                </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-failsafe-plugin</artifactId>
                 <version>2.18.1</version>
                 <executions>
                     <execution>
+                        <id>test-on-jvm6</id>
                         <goals>
                             <goal>integration-test</goal>
                             <goal>verify</goal>
                         </goals>
+                        <configuration>
+                            <excludes>
+                                <exclude>**/jdk7/**/</exclude>
+                            </excludes>
+                            <!-- AbstractPinpointPluginTestSuite needs this to resolve
+                                path of required jars -->
+                            <useSystemClassLoader>false</useSystemClassLoader>
+                            <failIfNoTests>true</failIfNoTests>
+                            <jvm>${env.JAVA_6_HOME}/bin/java</jvm>
+                        </configuration>
+                    </execution>
+                    <execution>
+                        <id>test-on-jvm7</id>
+                        <goals>
+                            <goal>integration-test</goal>
+                            <goal>verify</goal>
+                        </goals>
+                        <configuration>
+                            <includes>
+                                <include>**/jdk7/**/</include>
+                            </includes>
+                            <!-- AbstractPinpointPluginTestSuite needs this to resolve
+                                path of required jars -->
+                            <useSystemClassLoader>false</useSystemClassLoader>
+                            <failIfNoTests>true</failIfNoTests>
+                            <jvm>${env.JAVA_7_HOME}/bin/java</jvm>
+                        </configuration>
                     </execution>
                 </executions>
-                <configuration>
-                    <!-- AbstractPinpointPluginTestSuite needs this to resolve 
-                        path of required jars -->
-                    <useSystemClassLoader>false</useSystemClassLoader>
-                    <failIfNoTests>true</failIfNoTests>
-                </configuration>
             </plugin>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
